Start playing with graphics
When running qemu without graphical output, getting ENOENT when opening /dev/fb0
When running qemu with graphical output, opening /dev/fb0 succeeds, but then gets EBADF when naively writing data to /dev/fb0

Did not need to create /dev/fb0 as a special character device.  It seems that the kernel itself creates this when it detects an available display device.